Split up Cloud client
